Migrate func_w0613.py to new functional tests
