+++ tests/functional/u/unused/unused_argument.py
@@ -29,7 +29,7 @@ class Sub(Base):
         "overridden method, though don't use every argument"
         return aaa
 
-    def newmethod(self, aax, aay): # [unused-argument]
+    def newmethod(self, aax, aay):  # [unused-argument]
         "another method, warning for aay desired"
         return self, aax
 
@@ -46,3 +46,44 @@ def metadata_from_dict(key):
     used inside comprehension dict
     """
     return {key: str(value) for key, value in key.items()}
+
+# pylint: disable=R0903, print-statement, misplaced-future,wrong-import-position
+from __future__ import print_function
+
+
+def function(arg=1):  # [unused-argument]
+    """ignore arg"""
+
+
+class AAAA(object):
+    """dummy class"""
+
+    def method(self, arg):  # [unused-argument]
+        """dummy method"""
+        print(self)
+    def __init__(self, *unused_args, **unused_kwargs):
+        pass
+
+    @classmethod
+    def selected(cls, *args, **kwargs):  # [unused-argument, unused-argument]
+        """called by the registry when the vobject has been selected.
+        """
+        return cls
+
+    def using_inner_function(self, etype, size=1):
+        """return a fake result set for a particular entity type"""
+        rset = AAAA([('A',)]*size, '%s X' % etype,
+                    description=[(etype,)]*size)
+        def inner(row, col=0, etype=etype, req=self, rset=rset):
+            """inner using all its argument"""
+            # pylint: disable = E1103
+            return req.vreg.etype_class(etype)(req, rset, row, col)
+        # pylint: disable = W0201
+        rset.get_entity = inner
+
+class BBBB(object):
+    """dummy class"""
+
+    def __init__(self, arg):  # [unused-argument]
+        """Constructor with an extra parameter. Should raise a warning"""
+        self.spam = 1
